Pull out gateway test file
